Add header position check dead zone for overscrolling
This greatly reduces the chances of unintentionally bringing up the header in Safari when scrolling all the way down in a swipe that results in an overscroll effect.
Other methods were attempted, but this simple (yet not so perfect) one seems to do a far better job with reduced complexity.
Due to Safari mobile behaving differently in regards to how many pixels it reports scrolling up to at the end of the effect (usually not just landing on the relative 0 position of the bottom like how it is on Safari desktop), there is a far larger deadzone to cover. The number defined in this commit was found to be the maximum on the test device. It might greater on others.
Let it be known: I tried.