Base MobileSafari detection on CSS support and media query
`user-select` is not present in Safari in general (only the prefixed version -- of which is an alias on other browsers). `(hover: hover)` being false implies in this case that a touch screen is being used (I know -- on the iPhone 15 or later, or on an iPad, you can use a mouse or trackpad -- haven't tested if this might throw detection off).
Somewhat under assumption that this impacts iPads too. If not.. not really much other reliable indications to go by.

Change was made as using desktop mode will change the user agent, which means the dead zone will not be applied with that enabled. Besides, shouldn't be relying on user agent for feature detection in any case.